BR-4247 Removed vardefmanager storing vardef files to cache/modules dir. vardefmanager will always store vardefs to MetadataCache now. Removed noFilesystemMetadataCache flag. Removed unit test TrackerTest.php because it was depending on vardefs being in filesystem which is no longer true. Fixed AssignedToMeTest that was modifying global dictionary in wrong way.

+++ sugarcrm/include/MetaDataManager/MetaDataCache.php
@@ -21,6 +21,7 @@ use Sugarcrm\Sugarcrm\Logger\Factory as LoggerFactory;
 class MetaDataCache implements LoggerAwareInterface
 {
     use LoggerAwareTrait;
+    protected $logger;
 
     protected $db;
 
@@ -35,6 +36,7 @@ class MetaDataCache implements LoggerAwareInterface
     public function __construct(DBManager $db)
     {
         $this->db = $db;
+        $this->setLogger(LoggerFactory::getLogger('metadata'));
     }
 
     public function get($key)
@@ -67,10 +69,15 @@ class MetaDataCache implements LoggerAwareInterface
 
     public function getKeys()
     {
-        return $this->db
-            ->getConnection()
-            ->executeQuery('SELECT type FROM ' . static::$cacheTable)
-            ->fetchAll(\PDO::FETCH_COLUMN);
+        $ret = array();
+        if (!empty($this->db) && self::$isCacheEnabled) {
+                $ret[] = $this->db
+                    ->getConnection()
+                    ->executeQuery('SELECT type FROM ' . static::$cacheTable)
+                    ->fetchAll(\PDO::FETCH_COLUMN);
+        }
+
+        return $ret;
     }
 
     public function reset()
@@ -92,6 +99,26 @@ class MetaDataCache implements LoggerAwareInterface
     }
 
     /**
+     *
+     * Enables MetaDataCache
+     *
+     */
+    public static function enableCache()
+    {
+        self::$isCacheEnabled = true;
+    }
+
+    /**
+     *
+     * Disables MetaDataCache
+     *
+     */
+    public static function disableCache()
+    {
+        self::$isCacheEnabled = false;
+    }
+
+    /**
      * Used to cache metadata responses in the database
      *
      * @param String $key key for data stored in the cache table
@@ -102,7 +129,7 @@ class MetaDataCache implements LoggerAwareInterface
     {
         $result = null;
         //During install/setup, this function might get called before the DB is setup.
-        if (!empty($this->db)) {
+        if (!empty($this->db) && self::$isCacheEnabled) {
             $cacheResult = null;
             $row = $this->getLastModifiedByType($key);
             if (!empty($row['data'])) {
@@ -131,7 +158,7 @@ class MetaDataCache implements LoggerAwareInterface
      */
     protected function storeToCacheTable($key, $data)
     {
-        if (!empty($this->db)) {
+        if (!empty($this->db) && self::$isCacheEnabled) {
             try {
                 $encoded = base64_encode(gzdeflate(serialize($data)));
             } catch (Exception $e) {
@@ -232,6 +259,9 @@ class MetaDataCache implements LoggerAwareInterface
     }
 
     public function clearKeysLike($key) {
+        if (!self::$isCacheEnabled) {
+            return true;
+        }
         $qb = $this->db->getConnection()->createQueryBuilder();
         return $qb->delete(static::$cacheTable)
             ->where($qb->expr()->like('type', $qb->createPositionalParameter($key . '%')))
